+++ components/network/libemqtt/luat_lib_mqtt.c
@@ -793,7 +793,7 @@ static int l_mqtt_autoreconn(lua_State *L) {
 mqttc:publish("/luatos/123456", "123")
 */
 static int l_mqtt_publish(lua_State *L) {
-	uint16_t message_id ,payload_len= 0;
+	uint32_t message_id ,payload_len= 0;
 	luat_mqtt_ctrl_t * mqtt_ctrl = get_mqtt_ctrl(L);
 	const char * topic = luaL_checkstring(L, 2);
 	const char * payload = luaL_checklstring(L, 3, &payload_len);
